fixed controllers

+++ src/controllers/contacts.js
@@ -1,8 +1,9 @@
 import createError from 'http-errors';
 import * as contactsService from '../services/contacts.js';
+import mongoose from 'mongoose';
 
 export const getAllContacts = async (req, res, _next) => {
-  const contacts = await contactsService.getAll();
+  const contacts = await contactsService.getAllContacts();
   res.json({
     status: 200,
     message: 'Successfully found contacts!',
@@ -12,7 +13,10 @@ export const getAllContacts = async (req, res, _next) => {
 
 export const getContactById = async (req, res, next) => {
   const { contactId } = req.params;
-  const contact = await contactsService.getById(contactId);
+  if (!mongoose.Types.ObjectId.isValid(contactId)) {
+    return next(createError(404, 'Contact not found'));
+  }
+  const contact = await contactsService.getContactById(contactId);
   if (!contact) {
     const error = createError(404, 'Contact not found');
     return next(error);
@@ -46,6 +50,9 @@ export const createContact = async (req, res, _next) => {
 export const updateContact = async (req, res, next) => {
   const { contactId } = req.params;
   const updateData = req.body;
+  if (!mongoose.Types.ObjectId.isValid(contactId)) {
+    return next(createError(404, 'Contact not found'));
+  }
 
   try {
     const updatedContact = await contactsService.updateContact(contactId, updateData);
@@ -67,6 +74,10 @@ export const updateContact = async (req, res, next) => {
 
 export const deleteContact = async (req, res, next) => {
   const { contactId } = req.params;
+  
+  if (!mongoose.Types.ObjectId.isValid(contactId)) {
+    return next(createError(404, 'Contact not found'));
+  }
 
   try {
     const deletedContact = await contactsService.deleteContact(contactId);
